Converted integration test for ReportListParameters to unit tests
